requirements.txt
yfinance
pandas
streamlit
plotly
requests
matplotlib
python-dotenv
ta
schedule
pytz
